SAT-905: Update cray-import-config requirement version
This updates the base cray-import-config helm chart version
to ~0.2.7.

This change updates the helm chart version and docker image
version for the sat-cfs-install docker image and helm chart to 1.0.1

Finally, this change updates the sat_product_version string
with the assumption that the next sat product stream version will
be incremented accordingly.

Test Description: I built a test product stream with this version
of the helm chart and docker image, and I deployed on baldar. I
followed the installation instructions to create a CFS session
and ran the ansible plays to update the cray-sat-podman RPM.