SAT-930: Remove default versions from values.yaml
This commit removes ``/product_version`` from the Docker image
and removes ``CF_IMPORT_PRODUCT_VERSION`` from the Helm chart's
``values.yaml``, with the expectation that this value will be
supplied via the Loftsman manifest. This way, the SAT product
stream version does not have to be specified in the config import
Helm chart.

Second, this commit removes the default config_image tag, so
that the helm chart version is not as strictly tied to the Docker
image version.

For both of these changes, variable substitution is removed
from ``runBuildPrep.sh``.

This commit also defines version 1.0.2 with these changes.

Test Description: Build a release distribution with this helm chart/
docker image version and did a test install of it on thanos.