CRAYSAT-1177: Enable GPG checking RPMs on release branch
Enable GPG checking on RPMs on release branch builds. Only enable GPG
checking on RPMs, not on the repository metadata, which is not signed.
If this is not a release branch build, the RPMs will not be signed, so
disable GPG checks entirely on non-release-branch builds.

The `runBuildPrep.sh` script is used to control whether GPG checking is
enabled or disabled on the repos by substituting in the appropriate
values in the Ansible role's `defaults/main.yml`.

Test Description:
Ran runBuildPrep.sh locally and copied resulting Ansible role to a
system that has a master-branch build with unsigned RPMs (hela),
uninstalled the cray-sat-podman RPM on a master node and
updated the SAT layer of ncn-personalization to point at a branch with
the new content. Ensured that the cray-sat-podman RPM was installed.

Did the same but built locally on a branch that starts with "release/"
and tested on a system installed with a SAT 2.2 release branch build of
SAT (baldar). Confirmed that the RPM was installed, and GPG checking on
the packages was enabled on the repository.