CRAYSAT-1337: Remove SLE repos from SAT Ansible
Stop adding SLE repos in the sat-ncn Ansible role. These were only being
added to ensure that the `podman` and `podman-cni-config` packages upon
which the `cray-sat-podman` package depends could be installed. These
packages are installed in the NCN images by default, so this is not
necessary, and having these repositories here causes more problems than
it solves. In the future, SLE may be providing its own CFS layer that
adds these repositories on NCNs.

Test Description:
Not tested. Should be very low risk as we are just reducing the
iterations of the loop. For a full test that it resolves the problem, we
would want to try installing SAT on a system without SLE installed and
verify that this Ansible role succeeds.