CRAYSAT-1363: Update to CSM 1.2 base image and helm chart
Update the version of the base cf-gitea-import container image to
1.4.69, which is documented as the required version for compatibility
with CSM 1.2. Bump the sat-cfs-install container image minor version to
correspond with the base image minor version bump.

Update the version of the cray-import-config Helm chart on which this
chart depends to version 2.0.0, which is documented as the required
version for compatibility with CSM 1.2. Bump the helm chart major
version to match the base chart major version bump. Change the way the
image is specified in `values.yaml` to match the way the version 2.0.0
`cray-import-config` Helm chart expectes it. Note that similar changes
are required to the way the `config_image.tag` and `catalog.image`
values are passed in through the loftsman manifest in the
sat-product-stream repository.

Also upgrade the Helm chart to use Helm API version v2 and put the
dependency charts directly in the Chart.yaml since `requirements.yaml`
has been deprecated.

Test Description:
Pushed to GitHub, waited for CJE to build it. Then referenced the
resulting Docker image and Helm charts which were uploaded to the
sat-docker-unstable-local and sat-helm-unstable-local arti repo in a
correpsonding branch of hpc-sat-product-stream. Pushed that branch and
installed on mug with CSM 1.2. Still need to test on a CSM 1.0 system
for backwards compatibility.